api-gateway/requirements.txt
alembic==1.13.2
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
importlib_metadata==8.4.0
importlib_resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0
openai==1.45.0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pydantic==2.9.1
pydantic_core==2.23.3
PyJWT==2.9.0
pytest==8.3.3
pytest-dotenv==0.5.2
pytest-flask==1.3.0
pytest-mock==3.14.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
tomli==2.0.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
zipp==3.20.1
responses==0.23.3